requirements.txt
fastapi
uvicorn[standard]
openai>=1.40.0
langdetect>=1.0.9
PyYAML
